Rework NodeKind metadata:
NodeKind used to contain and index and 2 bits of metadata,
this was used to both differentiate NodeEvent tree nodes
from tree leaves and to mark skip/trivia tokens.

Since we've introduced word/keyword and error tokens, 2 bits have been
inadequate so now there are 3 bits of metadata: 1 bit for tree/leaf
and 2 bits for token variant.